+++ overlord_exe/src/main/kotlin/club/subjugated/overlord_exe/bots/timer_bot/TimerBotRecordService.kt
@@ -114,7 +114,7 @@ class TimerBotRecordService(
         val terms = if(record.isRandom) {
             "Duration revealed on accept"
         } else {
-            formatDuration(maxDurationAmount, form.maxUnit)
+            formatDuration(maxDurationAmount.toLong(), form.maxUnit)
         }
 
         // Issue the contract
