+++ overlord_exe/src/main/kotlin/club/subjugated/overlord_exe/util/TimeHelper.kt
@@ -2,14 +2,14 @@ package club.subjugated.overlord_exe.util
 
 import java.time.temporal.ChronoUnit
 
-fun formatDuration(amount: Int, unit: String): String {
+fun formatDuration(amount: Long, unit: String): String {
     val chronoUnit = ChronoUnit.valueOf(unit)
     val label = when (chronoUnit) {
-        ChronoUnit.SECONDS -> if (amount == 1) "second" else "seconds"
-        ChronoUnit.MINUTES -> if (amount == 1) "minute" else "minutes"
-        ChronoUnit.HOURS   -> if (amount == 1) "hour" else "hours"
-        ChronoUnit.DAYS    -> if (amount == 1) "day" else "days"
-        ChronoUnit.WEEKS -> if (amount == 1) "week" else "weeks"
+        ChronoUnit.SECONDS -> if (amount == 1L) "second" else "seconds"
+        ChronoUnit.MINUTES -> if (amount == 1L) "minute" else "minutes"
+        ChronoUnit.HOURS   -> if (amount == 1L) "hour" else "hours"
+        ChronoUnit.DAYS    -> if (amount == 1L) "day" else "days"
+        ChronoUnit.WEEKS -> if (amount == 1L) "week" else "weeks"
         else -> unit.toString().lowercase()
     }
     return "$amount $label"
